+++ marsworks/manifest.py
@@ -21,74 +21,43 @@ LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 SOFTWARE.
 """
+from __future__ import annotations
 
 from datetime import date, datetime
-from typing import Optional
+from typing import Optional, Mapping, Any
 
-from .origin.exceptions import BadContentError
-from .origin.internal_utils import repr_gen
+from .partialmanifest import PartialManifest
 
 __all__ = ("Manifest",)
 
 
-class Manifest:
+class Manifest(PartialManifest):
     """
     A class representing a `Manifest`.
 
     Attributes:
 
-        rover_id (int): ID of the rover.
-        name (str): Name of the Rover.
-        status (str): The Rover's mission status.
-        max_sol (int): The most recent Martian sol from which photos exist.
-        total_photos (int): Number of photos taken by that Rover.
-        cameras (dict): Cameras for which there are photos by that Rover on that sol.
-    """
+        rover_id (Optional[int]): ID of the rover.
+        rover_name (str): Name of the Rover.
+        status (Optional[str]): The Rover's mission status.
+        max_sol (Optional[int]): The most recent Martian sol from which photos exist.
+        total_photos (Optiona[int]): Number of photos taken by that Rover.
+        cameras (Mapping[str, str]): Cameras for which there are photos by that Rover on that sol.
+    """  # noqa: E501
 
     __slots__ = (
-        "_data",
-        "rover_id",
-        "name",
-        "status",
         "max_sol",
         "total_photos",
         "cameras",
     )
 
-    def __init__(self, data: dict) -> None:
-        self._data: dict = data
-        self.rover_id: Optional[int] = data.get("id")
-        self.name: Optional[str] = data.get("name")
-        self.status: Optional[str] = data.get("status")
+    def __init__(self, data: Mapping[Any, Any]) -> None:
+        super().__init__(data)
         self.max_sol: Optional[int] = data.get("max_sol")
         self.total_photos: Optional[int] = data.get("total_photos")
-        self.cameras: Optional[dict] = data.get("cameras")
+        self.cameras: Mapping[str, str] = data.get("cameras", {})
 
-    def __len__(self) -> int:
-        """
-        Returns:
-
-            length of internal dict of attributes. (Result of `len(obj)`)
-        """
-        return len(self._data)
-
-    def __repr__(self) -> str:
-        """
-        Returns:
-
-            Representation of Manifest. (Result of `repr(obj)`)
-        """
-        return repr_gen(__class__, self)
-
-    def __str__(self) -> Optional[str]:
-        """
-        Returns:
-
-            Name of the Rover. (Result of `str(obj)`)
-        """
-        return self.name
-
-    def __eq__(self, value) -> bool:
+    def __eq__(self, value: Any) -> bool:
         """
         Checks if two objects are same using `rover_id`.
 
@@ -97,36 +66,6 @@ class Manifest:
             Result of `obj == obj`.
         """
         return isinstance(value, self.__class__) and value.rover_id == self.rover_id
-
-    def __hash__(self) -> int:
-        """
-        Returns:
-
-            hash of the class. (Result of `hash(obj)`)
-        """
-        return hash(self.__class__)
-
-    @property
-    def launch_date(self) -> date:
-        """
-        The Rover's launch date from Earth.
-
-        Returns:
-
-            A [datetime.date](https://docs.python.org/3/library/datetime.html?highlight=datetime%20date#datetime.date) object.
-        """  # noqa: E501
-        return datetime.date(datetime.strptime(self._data["launch_date"], "%Y-%m-%d"))
-
-    @property
-    def landing_date(self) -> date:
-        """
-        The Rover's landing date on Mars.
-
-        Returns:
-
-            A [datetime.date](https://docs.python.org/3/library/datetime.html?highlight=datetime%20date#datetime.date) object.
-        """  # noqa: E501
-        return datetime.date(datetime.strptime(self._data["landing_date"], "%Y-%m-%d"))
 
     @property
     def max_date(self) -> date:
@@ -137,25 +76,6 @@ class Manifest:
 
             A [datetime.date](https://docs.python.org/3/library/datetime.html?highlight=datetime%20date#datetime.date) object.
         """  # noqa: E501
-        return datetime.date(datetime.strptime(self._data["max_date"], "%Y-%m-%d"))
-
-    def search_camera(self, camera: str) -> list:
-        """
-        Looks for the camera supplied.
-
-        Args:
-
-            camera: The camera to look for. (Must be in Upper case and short name. like: `PANCAM`)
-
-        Returns:
-
-            list of cameras with that name.
-        """  # noqa: E501
-        camera_data = self.cameras
-        if isinstance(camera_data, list):
-            try:
-                return [cam["name"] for cam in camera_data if camera == camera]
-            except KeyError:
-                raise BadContentError(content=camera_data) from None
-        else:
-            raise BadContentError(message=f"can't iterate over <{camera_data}>.")
+        return datetime.date(
+            datetime.strptime(self._rover_info["max_date"], "%Y-%m-%d")
+        )